requirements.txt
adal==1.2.4
aiofiles==0.6.0
aiohttp==3.7.4
argon2-cffi==20.1.0
async-timeout==3.0.1
attrs==20.1.0
azure-cognitiveservices-search-imagesearch==2.0.0
azure-common==1.1.25
backcall==0.2.0
bleach==3.1.5
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
cycler==0.10.0
cymem==2.0.3
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastai==2.0.6
fastbook==0.0.8
fastcore==1.0.0
fastprogress==1.0.0
fastscript==1.0.0
future==0.18.2
h11==0.12.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
msrest==0.6.18
msrestazure==0.6.4
multidict==5.1.0
murmurhash==1.0.2
nbconvert==5.6.1
nbdev==1.0.10
nbformat==5.0.7
networkx==2.5
notebook==6.1.3
numpy==1.19.1
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.86
six==1.15.0
spacy==2.3.2
srsly==1.0.2
starlette==0.14.2
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2021.3.5
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==5.0.0
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.13.4
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.1.0
